pip_requirements.txt
aiohttp~=3.7.4
discord.py~=1.7.3
ffmpeg-python~=0.2.0
python-dateutil~=2.8
python-dotenv~=0.19.0
pytube~=12.1
pytz~=2021.1
requests~=2.28
sqlalchemy~=2.0.0b3
